Support Coverage Types
There are three coverage types: Count, Significance, and EffectSize. The numbers associated with each bucket change depending on the coverage type.

Count: log10_sig is the largest significance in the bucket and effect is the
        effect with the largest absolute value
Significance: log10_sig is the largest significance in the bucket and effect
        is the effect size associated with that significance value
EffectSize: effect is the effect with the largest absolute value and log10_sig
        is the significance associated with that effect.